+++ runtime/commonMain/src/kotlinx/serialization/descriptors/SerialDescriptorBuilder.kt
@@ -2,9 +2,9 @@
  * Copyright 2017-2020 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package kotlinx.serialization
+package kotlinx.serialization.descriptors
 
-import kotlinx.serialization.descriptors.*
+import kotlinx.serialization.*
 import kotlinx.serialization.encoding.*
 import kotlinx.serialization.internal.*
